Moving the hashmaputils.m to the matlab server.
This script can be used on both, the client and the server side to transfer data. Hence the server package is a better place for it.
The resource folder was made a source folder so that the class loader finds the script during runtime using absolute paths.
Removing fileNameTrunk method from the RGG package to the MatlabUtilities. Adjust the dependent Classes accordingly