Created a new MATLAB client
The new client will have the possibility to run locally and remotely. So far the local client was implemented. The remote client will later be adapted to also use the MatlabController class.
The code was reorganised: There are two new classes MatlabTable and MatlabCode, that take a lot of code from the MatlabUtilities and the TableConverter.
The MatlabController class had some methods renamed. Redundant code was removed. Most importantly the way how the ArrayBlockingQueue for the MatlabProxies is handled was changed; now a new MATLAB session is created only when it is used. This has the advantage that the loading of the workflows is quicker, since it does not immediatly launch the MATLAB application at this moment.