Adapt the hashmaputils.m script
Removing all the standard output statements, since this is now handled by the MatlabCode class and called in the task methods of the MatlabClient.
Suppress output of dynamically generated MATLAB code.

Fixing a bug; When the we can’t checkout the StatisticsToolbox, we fall back on the map-type. The dataType variable was not changed accordingly.

Simplifying the code for type (class) checking.