Remove output type parameter from the MATLAB plot node
the parameter is not used anymore and was partially commented out. Since the dialog did not allow anymore to edit this option, the setting was never set. However during loading the node, the node model tried to load this setting never the less and then complained about the absence of the output type setting.