Change MatlabClient instantiation
The instantiation is now done in the last moment, meaning only when a node is executed. This way the workflow can load normally and possible connection exceptions occur only after loading the workflow.

The instantiation of the Remote host was now wrapped in a executor to impose a timeout. The cajo library has a veeery long timeout, because it is not possible to differentiate between queuing and actual timeout. Hence we bypass this, only during instantiation, with the executor. Experience shows, that if the network connection is there and the server is running, it connects quickly.